CCP-2496 Move SearchBar in Header components
