+++ libraries/ui-shared/Sheet/components/Header/components/SearchBar/style.js
